+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Changelog
 * Add new method :meth:`pybam.slice`.
 * Update :command:`bam-slice` command to use :meth:`pybam.slice` method.
 * Update :command:`ngs-bam2fq` and :command:`ngs-fq2bam` commands to allow users to run in local environment.
+* Update :command:`ngs-fq2bam` command to handle cases where input FASTQ does not have information on flowcell and barcode.
 
 0.32.0 (2022-04-02)
 -------------------